feature(kubernetes plugin): Adds support for a kuberentes plugin fixes #569 This commit also refactors the existing docker plugin so that the kubernetes plugin can gather metrics from the docker system.
