+++ .gitignore
@@ -4,3 +4,4 @@ tivan
 .idea
 *~
 *#
+vendor/
