+++ glide.yaml
@@ -1,0 +1,157 @@
+package: github.com/influxdata/telegraf
+import:
+- package: git.eclipse.org/gitroot/paho/org.eclipse.paho.mqtt.golang.git
+  version: dbd8d5c40a582eb9adacde36b47932b3a3ad0034
+- package: github.com/Shopify/sarama
+  version: d37c73f2b2bce85f7fa16b6a550d26c5372892ef
+- package: github.com/Sirupsen/logrus
+  version: f7f79f729e0fbe2fcc061db48a9ba0263f588252
+- package: github.com/amir/raidman
+  version: 6a8e089bbe32e6b907feae5ba688841974b3c339
+- package: github.com/armon/go-metrics
+  version: 345426c77237ece5dab0e1605c3e4b35c3f54757
+- package: github.com/aws/aws-sdk-go
+  version: 3ad0b07b44c22c21c734d1094981540b7a11e942
+  subpackages:
+  - aws
+  - aws/credentials
+  - aws/credentials/ec2rolecreds
+  - aws/ec2metadata
+  - aws/session
+  - service/cloudwatch
+  - service/kinesis
+- package: github.com/beorn7/perks
+  version: b965b613227fddccbfffe13eae360ed3fa822f8d
+- package: github.com/boltdb/bolt
+  version: 6465994716bf6400605746e79224cf1e7ed68725
+- package: github.com/cenkalti/backoff
+  version: 4dc77674aceaabba2c7e3da25d4c823edfb73f99
+- package: github.com/dancannon/gorethink
+  version: ff457cac6a529d9749d841a733d76e8305cba3c8
+- package: github.com/davecgh/go-spew
+  version: 5215b55f46b2b919f50a1df0eaa5886afe4e3b3d
+- package: github.com/eapache/go-resiliency
+  version: b86b1ec0dd4209a588dc1285cdd471e73525c0b3
+- package: github.com/eapache/queue
+  version: ded5959c0d4e360646dc9e9908cff48666781367
+- package: github.com/fsouza/go-dockerclient
+  version: 6fb38e6bb3d544d7eb5b55fd396cd4e6850802d8
+- package: github.com/go-ini/ini
+  version: afbd495e5aaea13597b5e14fe514ddeaa4d76fc3
+- package: github.com/go-sql-driver/mysql
+  version: 72ea5d0b32a04c67710bf63e97095d82aea5f352
+- package: github.com/gogo/protobuf
+  version: c57e439bad574c2e0877ff18d514badcfced004d
+- package: github.com/golang/protobuf
+  version: 2402d76f3d41f928c7902a765dfc872356dd3aad
+- package: github.com/golang/snappy
+  version: 723cc1e459b8eea2dea4583200fd60757d40097a
+- package: github.com/gonuts/go-shellquote
+  version: e842a11b24c6abfb3dd27af69a17f482e4b483c2
+- package: github.com/hailocab/go-hostpool
+  version: 50839ee41f32bfca8d03a183031aa634b2dc1c64
+- package: github.com/hashicorp/go-msgpack
+  version: fa3f63826f7c23912c15263591e65d54d080b458
+- package: github.com/hashicorp/raft
+  version: b95f335efee1992886864389183ebda0c0a5d0f6
+- package: github.com/hashicorp/raft-boltdb
+  version: d1e82c1ec3f15ee991f7cc7ffd5b67ff6f5bbaee
+- package: github.com/influxdata/config
+  version: bae7cb98197d842374d3b8403905924094930f24
+- package: github.com/influxdata/influxdb
+  version: 0e0f85a0c1fd1788ae4f9145531b02c539cfa5b5
+  subpackages:
+  - client/v2
+  - models
+  - services/graphite
+- package: github.com/influxdb/influxdb
+  version: 0e0f85a0c1fd1788ae4f9145531b02c539cfa5b5
+- package: github.com/jmespath/go-jmespath
+  version: c01cf91b011868172fdcd9f41838e80c9d716264
+- package: github.com/klauspost/crc32
+  version: 999f3125931f6557b991b2f8472172bdfa578d38
+- package: github.com/lib/pq
+  version: 8ad2b298cadd691a77015666a5372eae5dbfac8f
+- package: github.com/matttproud/golang_protobuf_extensions
+  version: d0c3fe89de86839aecf2e0579c40ba3bb336a453
+- package: github.com/mreiferson/go-snappystream
+  version: 028eae7ab5c4c9e2d1cb4c4ca1e53259bbe7e504
+- package: github.com/naoina/go-stringutil
+  version: 6b638e95a32d0c1131db0e7fe83775cbea4a0d0b
+- package: github.com/naoina/toml
+  version: 751171607256bb66e64c9f0220c00662420c38e9
+  subpackages:
+  - ast
+- package: github.com/nsqio/go-nsq
+  version: 2118015c120962edc5d03325c680daf3163a8b5f
+- package: github.com/pborman/uuid
+  version: dee7705ef7b324f27ceb85a121c61f2c2e8ce988
+- package: github.com/pmezard/go-difflib
+  version: 792786c7400a136282c1664665ae0a8db921c6c2
+- package: github.com/prometheus/client_golang
+  version: 67994f177195311c3ea3d4407ed0175e34a4256f
+  subpackages:
+  - prometheus
+- package: github.com/prometheus/client_model
+  version: fa8ad6fec33561be4280a8f0514318c79d7f6cb6
+- package: github.com/prometheus/common
+  version: 0a3005bb37bc411040083a55372e77c405f6464c
+  subpackages:
+  - expfmt
+  - model
+- package: github.com/prometheus/procfs
+  version: 406e5b7bfd8201a36e2bb5f7bdae0b03380c2ce8
+- package: github.com/samuel/go-zookeeper
+  version: 218e9c81c0dd8b3b18172b2bbfad92cc7d6db55f
+- package: github.com/shirou/gopsutil
+  version: 8850f58d7035653e1ab90711481954c8ca1b9813
+  subpackages:
+  - cpu
+  - disk
+  - host
+  - load
+  - mem
+  - net
+  - process
+- package: github.com/soniah/gosnmp
+  version: b1b4f885b12c5dcbd021c5cee1c904110de6db7d
+- package: github.com/streadway/amqp
+  version: b4f3ceab0337f013208d31348b578d83c0064744
+- package: github.com/stretchr/objx
+  version: 1a9d0bb9f541897e62256577b352fdbc1fb4fd94
+- package: github.com/stretchr/testify
+  version: f390dcf405f7b83c997eac1b06768bb9f44dec18
+  subpackages:
+  - assert
+  - mock
+- package: github.com/wvanbergen/kafka
+  version: 1a8639a45164fcc245d5c7b4bd3ccfbd1a0ffbf3
+  subpackages:
+  - consumergroup
+- package: github.com/wvanbergen/kazoo-go
+  version: 0f768712ae6f76454f987c3356177e138df258f8
+- package: golang.org/x/crypto
+  version: 3760e016850398b85094c4c99e955b8c3dea5711
+- package: golang.org/x/net
+  version: 72aa00c6241a8013dc9b040abb45f57edbe73945
+  subpackages:
+  - html/charset
+- package: golang.org/x/text
+  version: cf4986612c83df6c55578ba198316d1684a9a287
+- package: gopkg.in/dancannon/gorethink.v1
+  version: e2cef022d0495329dfb0635991de76efcab5cf50
+- package: gopkg.in/fatih/pool.v2
+  version: cba550ebf9bce999a02e963296d4bc7a486cb715
+- package: gopkg.in/mgo.v2
+  version: 03c9f3ee4c14c8e51ee521a6a7d0425658dd6f64
+  subpackages:
+  - bson
+- package: gopkg.in/yaml.v2
+  version: f7716cbe52baa25d2e9b0d0da546fcf909fc16b4
+- package: github.com/jchauncey/kubeclient
+  version: 52ad51a76eed34f58589952b20b4a1d5928cdcd2
+  subpackages:
+  - api
+  - client/unversioned
+- package: github.com/gorilla/mux
+- package: github.com/zensqlmonitor/go-mssqldb
